+++ src/training/graph_group_sync.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "optimizers/quantizer.h"
 #include "training/graph_group.h"
 #include "training/communicator.h"
 #include "training/exponential_smoothing.h"
@@ -23,6 +24,9 @@ class SyncGraphGroup : public GraphGroup, public ExponentialSmoothing {
   std::vector<Ptr<TensorAllocator>> paramsAllocs_; // [deviceIndex] we must hold a reference to the memory until this class dies
   // @TODO: move this nto ExponentialSmoothing, together with paramsAvg_?
 
+  // model quantizer
+  std::vector<Ptr<ModelQuantizer>> quantizers_;
+  
   // state for update()
   bool first_{ true };                           // gets interpreted and cleared by update()
   std::vector<Ptr<data::Batch>> pendingBatches_; // in case of dynamic MB-size scaling, we temporarly buffer up batches across update() calls until enough